refactor entrypoints and create exporter subcommand
