+++ cmd/seed_server/app/app.go
@@ -12,7 +12,6 @@ import (
 	"github.com/gardener/vpn2/pkg/openvpn"
 	"github.com/gardener/vpn2/pkg/pprof"
 	"github.com/gardener/vpn2/pkg/seed_server"
-	"github.com/gardener/vpn2/pkg/seed_server/openvpn_exporter"
 	"github.com/gardener/vpn2/pkg/utils"
 	"github.com/go-logr/logr"
 	"github.com/spf13/cobra"
@@ -50,6 +49,7 @@ func NewCommand() *cobra.Command {
 	flags := cmd.Flags()
 	verflag.AddFlags(flags)
 	cmd.AddCommand(firewallCommand())
+	cmd.AddCommand(exporterCommand())
 	cmd.PersistentFlags().BoolVar(&pprofEnabled, "enable-pprof", false, "enable pprof for profiling")
 	return cmd
 }
@@ -65,20 +65,6 @@ func run(ctx context.Context, log logr.Logger) error {
 		return err
 	}
 
-	log.Info("using openvpn network", "openVPNNetwork", v.OpenVPNNetwork)
-	openVPN, err := openvpn.NewServer(v)
-	if err != nil {
-		return fmt.Errorf("error creating openvpn server: %w", err)
-	}
-
-	if cfg.StatusPath != "" {
-		exporterConfig := openvpn_exporter.NewDefaultConfig()
-		exporterConfig.OpenvpnStatusPaths = cfg.StatusPath
-		exporterConfig.ListenAddress = fmt.Sprintf(":%d", metricsPort)
-		if err := openvpn_exporter.Start(log, exporterConfig); err != nil {
-			return fmt.Errorf("starting metrics exporter failed: %w", err)
-		}
-	}
-
-	return openVPN.Run(ctx)
+	log.Info("writing openvpn config file", "values", v)
+	return openvpn.WriteServerConfigFiles(v)
 }
